Regenerate site with a later pelican version
Whatever the old version of Pelican used was I can't remember. I haven't
worked on this in a long time and I've lost track of the old development
environment, so I've had to upgrade and the generated HTML has changed a
bit.

Mostly what that seems to mean is generated HTML for code blocks has
changed a bit, but I can't detect a visual difference. Also the
sub-section of the site and article is on is now added as a "<category>"
tag in the HTML.